added more error handling and assertions, now synchronizes all monsters instead of only the selected one

+++ HunterPie/Core/Game.cs
@@ -133,7 +133,13 @@ namespace HunterPie.Core
             if (Player.InPeaceZone) Time = null;
             else { Time = DateTime.UtcNow; }
 
-            synchandler.deleteSession();
+            if (UserSettings.PlayerConfig.HunterPie.Sync.Enabled)
+            {
+                if (synchandler.isPartyLeader && synchandler.partyExists())
+                {
+                    synchandler.deleteSession();
+                }
+            }
         }
 
         private void StartGameScanner()
